+++ app/pages/user/aspiration/list-page.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
 import { useAspiration } from '~/composables/api/useAspiration'
+import { useRouter } from 'vue-router'
 
 definePageMeta({
   middleware: 'auth',
@@ -9,6 +10,12 @@ definePageMeta({
 const { fetchAspirations } = useAspiration()
 const { data: response, pending, error, refresh } = await fetchAspirations()
 const listAspirasi = computed(() => response.value?.data || [])
+const router = useRouter()
+
+const goToDetail = (id: string) => {
+  console.log("ID yang diklik:", id)
+  router.push(`/user/aspiration/detail/${id}`)
+}
 </script>
 
 <template>
@@ -54,6 +61,7 @@ const listAspirasi = computed(() => response.value?.data || [])
         v-for="item in listAspirasi" 
         :key="item.id" 
         :report="item" 
+        @click="goToDetail(item.id)" class="cursor-pointer"
       />
     </div>
   </div>